requirements.txt
attrs==21.2.0
importlib-metadata==4.7.1
iniconfig==1.1.1
packaging==21.0
Pillow==8.3.1
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.4
selenium==3.141.0
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.26.6
zipp==3.5.0
